phase1-core-infrastructure/cloud/requirements.txt
flask==2.3.3
cassandra-driver==3.28.0
requests==2.31.0